feat: refactor safe_slack_api_call() to be a decorator
